+++ app/src/main/java/com/kingzcheung/xime/service/ImeKeyRouter.kt
@@ -1,0 +1,886 @@
+package com.kingzcheung.xime.service
+
+import android.util.Log
+import android.view.KeyEvent
+import android.view.inputmethod.EditorInfo
+import com.kingzcheung.xime.association.AssociationManager
+import com.kingzcheung.xime.keyboard.OverlayRoute
+import com.kingzcheung.xime.rime.resolveRimeCandidateIndex
+import com.kingzcheung.xime.settings.SettingsPreferences
+import com.kingzcheung.xime.ui.keyboard.KeyboardLayoutState
+import com.kingzcheung.xime.ui.keyboard.isT9Schema
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+
+/**
+ * 按键输入路由。
+ *
+ * 承载按键派发（handleKeyPress）、长按退格合并、候选选择/翻页、计算器候选等逻辑。
+ * 所有共享状态通过 service 引用访问（同模块 internal 成员）。
+ */
+internal class ImeKeyRouter(private val service: XimeInputMethodService) {
+    internal fun handleKeyPress(key: String, isShifted: Boolean) {
+        if (service.uiState.value.quickSendFormFocused) {
+            when (key) {
+                "enter" -> {
+                    val editText = QuickSendFormEditTextHolder.editText
+                    val text = editText?.text?.toString() ?: ""
+                    val s = service.uiState.value
+                    val editingId = s.quickSendEditingItemId
+                    if (text.isNotBlank()) {
+                        if (editingId != null) {
+                            service.keyboardViewModel.updateQuickSendItem(editingId, text)
+                        } else {
+                            service.keyboardViewModel.addQuickSendText(text)
+                        }
+                    }
+                    service.uiState.value = s.copy(showQuickSendForm = false, quickSendFormFocused = false, quickSendEditingItemId = null, quickSendEditingItemText = "")
+                    QuickSendFormEditTextHolder.editText = null
+                    service.keyboardViewModel.showOverlay(OverlayRoute.Clipboard(1))
+                    return
+                }
+                "delete" -> {
+                    val candState = service.candidateState.value
+                    val isComposing = candState.isComposing || candState.inputText.isNotEmpty()
+                    if (isComposing) {
+                        // Rime 有组合态 → 转发退格到 Rime 清空候选字母/联想词
+                        service.rimeEngine.processKey(0xff08, 0)
+                        val result = service.rimeEngine.getProcessResult(true)
+                        if (result.inputText.isEmpty()) {
+                            service.rimeEngine.clearComposition()
+                        }
+                        service.uiEventChannel.trySend {
+                            service.sessionController.updateUIWithResult(result)
+                        }
+                    } else {
+                        // 无组合态 → 直接操作 EditText 删除已上屏文字
+                        QuickSendFormEditTextHolder.editText?.let { et ->
+                            val start = et.selectionStart.coerceAtLeast(0)
+                            val end = et.selectionEnd.coerceAtLeast(start)
+                            if (start == end && start > 0) {
+                                et.text?.delete(start - 1, start)
+                                try { et.setSelection(start - 1) } catch (_: Exception) {}
+                            } else if (end > start) {
+                                et.text?.delete(start, end)
+                                try { et.setSelection(start) } catch (_: Exception) {}
+                            }
+                        }
+                    }
+                    return
+                }
+            }
+        }
+        // 长按退格以固定频率重复派发，走合并路径，避免 keyJobs 堆积导致候选栏抖动
+        if (key == "delete") {
+            handleDeleteKey()
+            return
+        }
+        val job = service.serviceScope.launch(service.keyProcessingDispatcher, start = CoroutineStart.LAZY) {
+            val state = service.uiState.value
+            val candState = service.candidateState.value
+            var needsUIUpdate = false
+            var pendingResult: com.kingzcheung.xime.rime.RimeProcessResult? = null
+            var committedText: String? = null
+            
+            when (key) {
+                "clear_composition" -> {
+                    service.calculatorEngine.clear()
+                    updateCalculatorCandidates()
+                    // 清空 partial commit 累积（与 clear_all 一致）：残留的已选词会在下一轮输入
+                    // 被 buildT9DisplayState 拼进 preedit。
+                    service.t9PartialCommitTexts.clear()
+                    service.rimeEngine.clearComposition()
+                    service.candidateState.value = service.candidateState.value.copy(
+                        candidates = emptyList(),
+                        candidateComments = emptyList(),
+                        associationCandidates = emptyList(),
+                        pendingEnglishText = "",
+                        isShowingRecentClipboard = false
+                    )
+                    needsUIUpdate = true
+                }
+                "clear_all" -> {
+                    service.calculatorEngine.clear()
+                    updateCalculatorCandidates()
+                    // 撤回内容记录（供 undo_clear 恢复）：
+                    // 显示在输入框模式：getTextBeforeCursor 返回的 inputFieldText 已包含全部 preedit
+                    //（composing 区），candState.inputText 与它是同一份内容，再拼接会重复
+                    //（如 "ji hua"+"ji hua"="ji huaji hua"）；候选栏模式输入框只有已上屏文本，
+                    // 才需补上候选栏中的编码 candState.inputText。
+                    val codeInInputBox = SettingsPreferences.getInputTextLocation(service) ==
+                        SettingsPreferences.INPUT_TEXT_INPUT_BOX
+                    val inputFieldText = withContext(Dispatchers.Main) {
+                        service.currentInputConnection?.getTextBeforeCursor(XimeInputMethodService.SAFE_TEXT_LIMIT, 0)?.toString() ?: ""
+                    }
+                    service.lastClearedText = if (codeInInputBox) inputFieldText else inputFieldText + candState.inputText
+                    // 清空 partial commit 累积：否则残留的已选词（如右选"几乎"）会在下一轮输入
+                    // 被 buildT9DisplayState 拼进 preedit（"几乎ji hua"）。
+                    service.t9PartialCommitTexts.clear()
+                    service.rimeEngine.clearComposition()
+                    service.candidateState.value = service.candidateState.value.copy(
+                        candidates = emptyList(),
+                        candidateComments = emptyList(),
+                        associationCandidates = emptyList(),
+                        pendingEnglishText = "",
+                        inputText = "",
+                        isComposing = false,
+                        isShowingRecentClipboard = false
+                    )
+                    withContext(Dispatchers.Main) {
+                        service.currentInputConnection?.let {
+                            service.endComposingInputBox()
+                            // 删除输入框中所有文字
+                            val textLen = inputFieldText.length
+                            if (textLen > 0) {
+                                it.deleteSurroundingText(textLen, 0)
+                            }
+                        }
+                    }
+                    needsUIUpdate = true
+                }
+                "undo_clear" -> {
+                    val text = service.lastClearedText
+                    if (text.isNotEmpty()) {
+                        service.lastClearedText = ""
+                        withContext(Dispatchers.Main) {
+                            val ic = service.currentInputConnection
+                            if (ic != null) {
+                                ic.commitText(text, text.length)
+                            }
+                        }
+                    }
+                    needsUIUpdate = true
+                }
+                "enter" -> {
+                    service.calculatorEngine.clear()
+                    updateCalculatorCandidates()
+                    if (candState.isComposing) {
+                        // T9 模式提交完整预编辑（含 partial commit 累积），非 T9 模式用 RIME input。
+                        val isT9 = isT9Schema(state.currentSchemaId)
+                        val input = if (isT9 && candState.preeditText.isNotEmpty()) {
+                            candState.preeditText
+                        } else {
+                            service.rimeEngine.getInput()
+                        }
+                        if (input.isNotEmpty()) {
+                            withContext(Dispatchers.Main) { service.commitText(input) }
+                        }
+                        if (isT9) {
+                            // 内联 CLEAR_ALL 逻辑：同步清空，避免异步 postRimeJob 延迟导致后续 backspace 拿到旧状态。
+                            service.t9PartialCommitTexts.clear()
+                            service.rimeEngine.setInput("")
+                            service.rimeEngine.clearComposition()
+                        } else {
+                            service.rimeEngine.clearComposition()
+                        }
+                        withContext(Dispatchers.Main) { service.endComposingInputBox() }
+                        needsUIUpdate = true
+                    } else {
+                        service.rimeEngine.clearComposition()
+                        withContext(Dispatchers.Main) {
+                            val imeOptions = service.currentInputEditorInfo?.imeOptions ?: 0
+                            val action = imeOptions and EditorInfo.IME_MASK_ACTION
+                            val noEnterAction = imeOptions and EditorInfo.IME_FLAG_NO_ENTER_ACTION != 0
+                            when {
+                                noEnterAction -> service.sendDownUpKeyEvents(KeyEvent.KEYCODE_ENTER)
+                                action == EditorInfo.IME_ACTION_GO ||
+                                action == EditorInfo.IME_ACTION_SEARCH ||
+                                action == EditorInfo.IME_ACTION_SEND ||
+                                action == EditorInfo.IME_ACTION_NEXT ||
+                                action == EditorInfo.IME_ACTION_DONE ->
+                                    service.currentInputConnection?.performEditorAction(action)
+                                else -> service.sendDownUpKeyEvents(KeyEvent.KEYCODE_ENTER)
+                            }
+                        }
+                    }
+                    withContext(Dispatchers.Main) {
+                        service.candidateState.value = service.candidateState.value.copy(
+                            inputText = "",
+                            pendingEnglishText = "",
+                            candidates = emptyList(),
+                            candidateComments = emptyList(),
+                            associationCandidates = emptyList(),
+                            isComposing = false
+                        )
+                        if (isT9Schema(state.currentSchemaId)) {
+                            service.uiState.value = service.uiState.value.copy(
+                                t9ResetSignal = service.uiState.value.t9ResetSignal + 1,
+                                t9RightCandidateSelectedCount = 0,
+                                t9SelectedCandidatePinyin = ""
+                            )
+                        }
+                    }
+                }
+                "space" -> {
+                    val pendingEnglish = candState.pendingEnglishText
+
+                    if (pendingEnglish.isNotEmpty()) {
+                        withContext(Dispatchers.Main) {
+                            service.commitText(pendingEnglish + " ")
+                            service.candidateState.value = service.candidateState.value.copy(
+                                pendingEnglishText = "",
+                                associationCandidates = emptyList()
+                            )
+                        }
+                    } else if (candState.isComposing) {
+                        if (candState.candidates.isNotEmpty()) {
+                            selectCandidateAsync(0)
+                        } else {
+                            val input = candState.inputText
+                            if (input.isNotEmpty()) {
+                                withContext(Dispatchers.Main) {
+                                    service.commitText(input)
+                                }
+                                service.rimeEngine.clearComposition()
+                                // T9模式：清空partialCommit累积文本，避免下一轮输入
+                                // preedit中残留上一轮的提交内容（如"看"→下一轮"看jihua"）
+                                if (isT9Schema(state.currentSchemaId)) {
+                                    withContext(Dispatchers.Main) {
+                                        service.t9PartialCommitTexts.clear()
+                                        service.uiState.value = service.uiState.value.copy(
+                                            t9ResetSignal = service.uiState.value.t9ResetSignal + 1,
+                                            t9RightCandidateSelectedCount = 0,
+                                            t9SelectedCandidatePinyin = ""
+                                        )
+                                    }
+                                }
+                                needsUIUpdate = true
+                            }
+                        }
+                    } else {
+                        withContext(Dispatchers.Main) {
+                            service.commitText(" ")
+                        }
+                    }
+                }
+                "word_separator" -> {
+                    if (candState.isComposing || candState.inputText.isNotEmpty()) {
+                        val result = service.rimeEngine.processKeyAndGetResult(0x27, 0)
+                        if (result.processed) {
+                            service.uiEventChannel.trySend {
+                                service.sessionController.updateUIWithResult(result)
+                            }
+                        } else {
+                            needsUIUpdate = true
+                        }
+                    } else {
+                        needsUIUpdate = true
+                    }
+                }
+                "shift" -> {
+                }
+                "mode_change" -> {
+                }
+                "ime_switch" -> {
+                    withContext(Dispatchers.Main) {
+                        service.schemaController.switchInputMethod()
+                    }
+                }
+                "abc" -> {
+                    service.calculatorEngine.clear()
+                    updateCalculatorCandidates()
+                }
+                "number", "common_symbol" -> {
+                    // Number/CommonSymbol 内部切换由 KeyboardView 的 key handler 处理
+                }
+                "emoji" -> {
+                    withContext(Dispatchers.Main) {
+                        service.commitText("😊")
+                    }
+                }
+                else -> {
+                    val isNumberKeyboard = service.keyboardViewModel.keyboardState.value is com.kingzcheung.xime.ui.keyboard.KeyboardLayoutState.Number
+                    val isCommonSymbolKeyboard = service.keyboardViewModel.keyboardState.value is com.kingzcheung.xime.ui.keyboard.KeyboardLayoutState.CommonSymbol
+
+                    val routeResult = com.kingzcheung.xime.calculator.routeCalculatorKey(
+                        key = key,
+                        isNumberKeyboard = isNumberKeyboard,
+                        isCommonSymbolKeyboard = isCommonSymbolKeyboard,
+                        calculatorEngine = service.calculatorEngine,
+                    )
+                    if (routeResult is com.kingzcheung.xime.calculator.CalculatorRouteResult.Handled) {
+                        withContext(Dispatchers.Main) { service.commitText(routeResult.commitText) }
+                        if (isNumberKeyboard) updateCalculatorCandidates()
+                        needsUIUpdate = true
+                        return@launch
+                    }
+
+                    val pendingEnglish = candState.pendingEnglishText
+                    
+                    // 非计算器键清除计算器状态
+                    if (!key.matches(Regex("[0-9]")) && key !in listOf("+", "-", "*", "/", ".")) {
+                        if (service.calculatorEngine.isActive() || service.calculatorEngine.getCandidate() != null) {
+                            service.calculatorEngine.clear()
+                            updateCalculatorCandidates()
+                        }
+                    }
+                    
+                    // 计算器模式：追踪数字、运算符和小数点
+                    if (key.matches(Regex("[0-9]")) || key in listOf("+", "-", "*", "/", ".")) {
+                        if (key.matches(Regex("[0-9]")) || key == ".") {
+                            service.calculatorEngine.handleDigit(key)
+                        } else {
+                            service.calculatorEngine.handleOperator(key)
+                        }
+                        updateCalculatorCandidates()
+                    }
+                    
+                    // 所有按键统一经过 Rime 引擎
+                    // 字母键不进入此分支（即使 pendingEnglish 非空），需要继续积累编码
+                    if (pendingEnglish.isNotEmpty() && !key.matches(Regex("[a-zA-Z]"))) {
+                        val finalKey = key
+                        withContext(Dispatchers.Main) {
+                            service.commitText(pendingEnglish + finalKey)
+                            service.candidateState.value = service.candidateState.value.copy(
+                                pendingEnglishText = "",
+                                associationCandidates = emptyList()
+                            )
+                        }
+                    } else {
+                        val isChinese = !state.isAsciiMode
+                        val char = key
+                        val keyCode = key.lowercase()[0].code
+                        val mask = if (isShifted) KeyEvent.META_SHIFT_ON else 0
+                        val isLetter = key.matches(Regex("[a-zA-Z]"))
+                        val isShiftedChinese = isShifted && isChinese && isLetter
+
+                        // Shifted non-letter keys: send character code to Rime (like soft keyboard does),
+                        // avoiding Rime misinterpreting physical keycodes as internal actions.
+                        if (isShifted && !isLetter) {
+                            if (char.length == 1) {
+                                val charCode = char[0].code
+                                val processed = service.rimeEngine.processKey(charCode, 0)
+                                if (processed) {
+                                    val result = service.rimeEngine.getProcessResult(processed)
+                                    service.uiEventChannel.trySend {
+                                        if (result.committedText.isNotEmpty()) service.commitText(result.committedText)
+                                        service.sessionController.updateUIWithResult(result)
+                                        if (service.calculatorEngine.isActive()) updateCalculatorCandidates()
+                                    }
+                                } else {
+                                    committedText = char
+                                    needsUIUpdate = true
+                                }
+                            } else {
+                                committedText = char
+                                needsUIUpdate = true
+                            }
+                        } else {
+                            val result = service.rimeEngine.processKeyAndGetResult(keyCode, mask)
+                            if (result.processed) {
+                                if (isShiftedChinese && result.committedText != char) {
+                                    service.rimeEngine.clearComposition()
+                                    committedText = char
+                                    needsUIUpdate = true
+                                } else {
+                                    val committed = result.committedText
+                                    if (state.isAsciiMode && committed.isNotEmpty() && result.inputText.isEmpty() && result.candidates.isEmpty()) {
+                                        val current = candState.pendingEnglishText
+                                        val newPending = current + committed
+                                        service.candidateState.value = service.candidateState.value.copy(pendingEnglishText = newPending)
+                                        withContext(Dispatchers.Main) {
+                                            service.currentInputConnection?.setComposingText(newPending, 1)
+                                        }
+                                        service.uiEventChannel.trySend {
+                                            service.sessionController.updateUIWithResult(result)
+                                            if (service.calculatorEngine.isActive()) updateCalculatorCandidates()
+                                        }
+                                    } else {
+                                        service.uiEventChannel.trySend {
+                                            if (committed.isNotEmpty()) service.commitText(committed)
+                                            service.sessionController.updateUIWithResult(result)
+                                            if (service.calculatorEngine.isActive()) updateCalculatorCandidates()
+                                        }
+                                    }
+                                }
+                            } else {
+                                val isAscii = state.isAsciiMode
+                                if (!candState.isComposing || isShiftedChinese) {
+                                                    if (isAscii) {
+                                                        val charToCommit = if (isShifted) char.uppercase() else char.lowercase()
+                                                        val currentPending = candState.pendingEnglishText
+                                                        val newPending = currentPending + charToCommit
+                                                        // 用 setComposingText 建立 composing region，选关联候选时 service.commitText 自然替换
+                                                        withContext(Dispatchers.Main) {
+                                                            service.currentInputConnection?.setComposingText(newPending, 1)
+                                                        }
+                                                        service.candidateState.value = service.candidateState.value.copy(
+                                                            pendingEnglishText = newPending,
+                                                            associationCandidates = emptyList()
+                                                        )
+                                                        needsUIUpdate = true
+                                    } else {
+                                        committedText = char
+                                        needsUIUpdate = true
+                                    }
+                                } else {
+                                    val candidateText = if (service.rimeEngine.selectCandidate(0)) {
+                                        service.rimeEngine.commit()
+                                    } else {
+                                        ""
+                                    }
+                                    committedText = candidateText + char
+                                    needsUIUpdate = true
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            
+            if (needsUIUpdate) {
+                val result = pendingResult
+                val textToCommit = committedText
+                if (result != null) {
+                    service.uiEventChannel.trySend {
+                        if (textToCommit != null) {
+                            service.commitText(textToCommit)
+                        }
+                        service.sessionController.updateUIWithResult(result)
+                        if (service.calculatorEngine.isActive()) {
+                            updateCalculatorCandidates()
+                        }
+                    }
+                } else {
+                    val capturedInputText = service.rimeEngine.getInput()
+                    val capturedCandidates = service.rimeEngine.getCandidatesWithComments()
+                    val capturedIsAscii = service.rimeEngine.isAsciiMode()
+                    val capturedHasNext = service.rimeEngine.hasNextPage()
+                    val capturedHasPrev = service.rimeEngine.hasPrevPage()
+                    service.uiEventChannel.trySend {
+                        if (textToCommit != null) {
+                            service.commitText(textToCommit)
+                        }
+                        val pendingEnglish = service.candidateState.value.pendingEnglishText
+                        val (filteredTexts, filteredComments) = if (capturedIsAscii) {
+                            val filtered = capturedCandidates.filterNot { candidate ->
+                                candidate.text.any { it.code in 0x4E00..0x9FFF }
+                            }
+                            filtered.map { it.text } to filtered.map { it.comment }
+                        } else {
+                            capturedCandidates.map { it.text } to capturedCandidates.map { it.comment }
+                        }
+                        service.candidateState.value = service.candidateState.value.copy(
+                            inputText = capturedInputText,
+                            candidates = filteredTexts,
+                            candidateComments = filteredComments,
+                            isComposing = capturedInputText.isNotEmpty(),
+                            associationCandidates = if ((capturedIsAscii || !service.isChineseMode) && pendingEnglish.isEmpty()) emptyList() else service.candidateState.value.associationCandidates,
+                            isShowingRecentClipboard = false,
+                            hasNextPage = capturedHasNext,
+                            hasPrevPage = capturedHasPrev
+                        )
+                        service.uiState.value = service.uiState.value.copy(isAsciiMode = capturedIsAscii)
+                        if (pendingEnglish.isNotEmpty()) {
+                            service.serviceScope.launch {
+                                val candidates = service.predictionManager.getEnglishAssociations(pendingEnglish, PredictionManager.MAX_ASSOCIATION_COUNT)
+                                withContext(Dispatchers.Main) {
+                                    service.candidateState.value = service.candidateState.value.copy(associationCandidates = candidates)
+                                }
+                            }
+                        }
+                        if (service.calculatorEngine.isActive()) {
+                            updateCalculatorCandidates()
+                        }
+                    }
+                }
+            }
+        }
+        service.keyJobs.trySend(job)
+    }
+
+    /**
+     * 退格键合并入口（主线程调用）。
+     *
+     * 无退格任务在执行时启动一个；已有任务在执行/排队时只累加 [service.pendingDeleteCount]，
+     * 由执行中的任务完成后顺带排空。这样长按退格不会让 keyJobs 无限堆积，
+     * 候选栏 UI 更新保持平滑，抬手后最多多删 1~2 个字符。
+     */
+    internal fun handleDeleteKey() {
+        val shouldLaunch = synchronized(service.deleteCoalesceLock) {
+            if (service.deleteJobActive) {
+                service.pendingDeleteCount++
+                false
+            } else {
+                service.deleteJobActive = true
+                true
+            }
+        }
+        if (!shouldLaunch) return
+        launchDeleteJob()
+    }
+
+    /**
+     * 启动一个退格 job（keyJobs FIFO 保序）。完成后若仍有累积的退格请求，
+     * 通过 [maybeScheduleFollowUp] 排入下一个 job。这样 keyJobs 中至多有一个
+     * 退格 job 在执行/排队：既避免长按退格（~80ms 重复）超过 rime 退格吞吐时
+     * 把 keyJobs 堆满、导致候选栏 UI 以"迟到的跳帧"方式刷新（一闪一闪），
+     * 也保持与其它按键的相对顺序——合并的退格会在夹在中间的字母键之后执行。
+     */
+    internal fun launchDeleteJob() {
+        val job = service.serviceScope.launch(service.keyProcessingDispatcher, start = CoroutineStart.LAZY) {
+            try {
+                processDeleteKey()
+            } catch (t: Throwable) {
+                Log.e(XimeInputMethodService.TAG, "processDeleteKey failed", t)
+            } finally {
+                maybeScheduleFollowUp()
+            }
+        }
+        service.keyJobs.trySend(job)
+    }
+
+    /** 若长按退格期间有累积的请求，排入下一个退格 job；否则结束合并状态。 */
+    internal fun maybeScheduleFollowUp() {
+        val shouldLaunch = synchronized(service.deleteCoalesceLock) {
+            if (service.pendingDeleteCount == 0) {
+                service.deleteJobActive = false
+                false
+            } else {
+                service.pendingDeleteCount--
+                true
+            }
+        }
+        if (shouldLaunch) {
+            launchDeleteJob()
+        }
+    }
+
+    /** 单次退格处理（service.keyProcessingDispatcher 上执行）。 */
+    internal suspend fun processDeleteKey() {
+        val candState = service.candidateState.value
+        // 退格改变输入上下文：使在途的联想预测结果失效，防止过期结果迟到回填
+        // associationCandidates，导致长按退格删除时候选栏在"联想词↔空"之间闪动。
+        service.predictionManager.invalidatePendingPredictions()
+        // 计算器模式：追踪退格
+        service.calculatorEngine.handleDelete()
+        updateCalculatorCandidates()
+
+        // 数字/符号键盘：直接发送系统退格，不经过 Rime
+        // 防止 T9 残留状态被 Rime 退格修改导致 UI 不一致
+        val layoutState = service.keyboardViewModel.keyboardState.value
+        if (layoutState is KeyboardLayoutState.Number || layoutState is KeyboardLayoutState.Symbol) {
+            withContext(Dispatchers.Main) {
+                service.sendDownUpKeyEvents(KeyEvent.KEYCODE_DEL)
+            }
+        } else when {
+            // 1. 英文待处理文本：逐个删除字符，重新加载联想
+            candState.pendingEnglishText.isNotEmpty() -> {
+                val pendingLen = candState.pendingEnglishText.length
+                if (pendingLen > 1) {
+                    val newPending = candState.pendingEnglishText.dropLast(1)
+                    withContext(Dispatchers.Main) {
+                        service.sendDownUpKeyEvents(KeyEvent.KEYCODE_DEL)
+                        service.candidateState.value = service.candidateState.value.copy(
+                            pendingEnglishText = newPending,
+                            candidates = emptyList(),
+                            candidateComments = emptyList(),
+                            associationCandidates = emptyList()
+                        )
+                    }
+                    service.serviceScope.launch {
+                        val candidates = service.predictionManager.getEnglishAssociations(newPending, PredictionManager.MAX_ASSOCIATION_COUNT)
+                        withContext(Dispatchers.Main) {
+                            service.candidateState.value = service.candidateState.value.copy(associationCandidates = candidates)
+                        }
+                    }
+                } else {
+                    withContext(Dispatchers.Main) {
+                        service.sendDownUpKeyEvents(KeyEvent.KEYCODE_DEL)
+                        service.candidateState.value = service.candidateState.value.copy(
+                            pendingEnglishText = "",
+                            candidates = emptyList(),
+                            candidateComments = emptyList(),
+                            associationCandidates = emptyList(),
+                            isShowingRecentClipboard = false
+                        )
+                    }
+                }
+            }
+
+            // 2. Rime 编码中：让 Rime 处理退格，更新候选
+            candState.isComposing || candState.inputText.isNotEmpty() -> {
+                service.rimeEngine.processKey(0xff08, 0)
+                val result = service.rimeEngine.getProcessResult(true)
+                if (result.inputText.isEmpty()) {
+                    service.rimeEngine.clearComposition()
+                    // T9 部分提交：剩余编码删完后，已上屏/ composing 的部分候选词无法用
+                    // RIME 退格删除，会一直卡在候选栏。这里撤销最近一次部分提交：
+                    // 清空 composing 区域（或删除上屏文本）并从累积列表移除。
+                    if (service.t9PartialCommitTexts.isNotEmpty()) {
+                        val len = service.t9PartialCommitTexts.last().length
+                        withContext(Dispatchers.Main) {
+                            if (SettingsPreferences.getInputTextLocation(service)
+                                == SettingsPreferences.INPUT_TEXT_INPUT_BOX) {
+                                service.endComposingInputBox()
+                            } else {
+                                service.currentInputConnection?.deleteSurroundingText(len, 0)
+                            }
+                        }
+                        service.t9PartialCommitTexts.removeLastOrNull()
+                    }
+                }
+                service.uiEventChannel.trySend {
+                    service.sessionController.updateUIWithResult(result)
+                    if (service.calculatorEngine.isActive()) updateCalculatorCandidates()
+                }
+            }
+
+            // 3. 联想词或剪贴板：仅清空候选栏，不回删已上屏字符
+            candState.associationCandidates.isNotEmpty() || candState.isShowingRecentClipboard -> {
+                service.candidateState.value = service.candidateState.value.copy(
+                    candidates = emptyList(),
+                    candidateComments = emptyList(),
+                    associationCandidates = emptyList(),
+                    isShowingRecentClipboard = false
+                )
+            }
+
+            // 4. 无候选也无编码：直接回删已上屏文本
+            else -> {
+                service.predictionManager.deleteLastChar()
+
+                withContext(Dispatchers.Main) {
+                    service.sendDownUpKeyEvents(KeyEvent.KEYCODE_DEL)
+                }
+
+                service.candidateState.value = service.candidateState.value.copy(
+                    candidates = emptyList(),
+                    candidateComments = emptyList(),
+                    associationCandidates = emptyList(),
+                    isShowingRecentClipboard = false
+                )
+            }
+        }
+    }
+
+    /**
+     * Posts a rime operation to [keyJobs] for sequential execution.
+     * Ensures no interleaving with key processing.
+     */
+    internal fun postRimeJob(block: suspend CoroutineScope.() -> Unit) {
+        val job = service.serviceScope.launch(service.keyProcessingDispatcher, start = CoroutineStart.LAZY) {
+            block()
+        }
+        service.keyJobs.trySend(job)
+    }
+
+    suspend fun selectCandidateAsync(index: Int) {
+        val selectedCandidate = if (index < service.candidateState.value.candidates.size) {
+            service.candidateState.value.candidates[index]
+        } else null
+
+        val isT9 = isT9Schema(service.uiState.value.currentSchemaId)
+        val candidatePinyin = if (isT9 && index < service.candidateState.value.candidateComments.size) {
+            service.candidateState.value.candidateComments[index]
+        } else null
+
+        // 在 RIME 真正 select/commit 之前，先同步通知 T9 控制器消费数字。
+        // 控制器返回 true 表示输入序列已被该候选词完整消费，服务层应视为 full commit。
+        val candidateTextLength = selectedCandidate?.length ?: 0
+        val fullyConsumed = if (isT9) {
+            service.keyboardCallbacks?.onT9RightCandidateWillBeSelected?.invoke(candidatePinyin, candidateTextLength) ?: false
+        } else {
+            false
+        }
+
+        // T9：跳过 service.rimeEngine.selectCandidate，消费已由 T9 处理器（t9_processor）
+        // 独立完成。selectCandidate 会遗留 [confirmed, phony] 残留 composition 状态，
+        // 导致后续 forceSendToRime 的 setInput 无法正常重建候选项（对齐 main 分支）。
+        // 非 T9 才调用 selectCandidate，并用 resolveRimeCandidateIndex 修正候选 index，
+        // 避免 UI 候选被过滤/重排后上屏错词。
+        val selectRimeOk = if (isT9) {
+            true
+        } else {
+            val rimeIndex = if (selectedCandidate != null) {
+                resolveRimeCandidateIndex(index, selectedCandidate, service.rimeEngine.getCandidates().toList())
+            } else {
+                index
+            }
+            service.rimeEngine.selectCandidate(rimeIndex)
+        }
+        if (!selectRimeOk) return
+
+        // T9：跳过 service.rimeEngine.commit()，用用户点选的 selectedCandidate 作为权威上屏文本。
+        val committedText = if (isT9) "" else service.rimeEngine.commit()
+        // T9 模式下 fullyConsumed 是判断 full/partial commit 的唯一权威：
+        // 当控制器明确 partial commit 时，即使 RIME commit() 返回非空文本
+        // （RIME 内部做了 partial commit），也不应走 full commit 路径。
+        val isFullCommit = if (isT9) {
+            fullyConsumed && selectedCandidate != null
+        } else {
+            committedText.isNotEmpty()
+        }
+        if (isFullCommit) {
+            if (SettingsPreferences.isSmartPredictionEnabled(service) && selectedCandidate != null && AssociationManager.isInitialized()) {
+                if (service.predictionManager.lastCommittedText.isNotEmpty()) {
+                    val lastChar = service.predictionManager.lastCommittedText.last().toString()
+                    service.predictionManager.recordInputPair(lastChar, selectedCandidate)
+                }
+            }
+            // T9 full commit：以用户点选的候选词文本为权威上屏文本；
+            // 非 T9 仍用 RIME committedText（可能含简繁转换等处理）。
+            val textToMerge = if (isT9 && selectedCandidate != null) {
+                selectedCandidate
+            } else if (committedText.isNotEmpty()) {
+                committedText
+            } else {
+                selectedCandidate!!
+            }
+            // T9 full commit：partial commit 累积文本（未单独上屏，只存在于
+            // service.t9PartialCommitTexts 的 composing 区）与本次上屏文本是独立词，必须拼接，
+            // 不能去重（mergePartialCommitText 仅用于显示，会吞掉重复的词）。
+            // 但若点击的是 partial 词本身（RIME 无菜单、候选栏显示的就是 partial 词，
+            // 无拼音注释 candidatePinyin==null），则只提交累积的 partial 文本，避免重复。
+            val fullCommitText = if (isT9 && service.t9PartialCommitTexts.isNotEmpty()) {
+                if (candidatePinyin == null) {
+                    service.t9PartialCommitTexts.joinToString("")
+                } else {
+                    service.t9PartialCommitTexts.joinToString("") + textToMerge
+                }
+            } else {
+                textToMerge
+            }
+            withContext(Dispatchers.Main) {
+                service.commitText(fullCommitText)
+                service.t9PartialCommitTexts.clear()
+                service.candidateState.value = service.candidateState.value.copy(
+                    inputText = "",
+                    candidates = emptyList(),
+                    candidateComments = emptyList(),
+                    isComposing = false,
+                    hasNextPage = false,
+                    hasPrevPage = false,
+                    isShowingRecentClipboard = false
+                )
+                service.uiState.value = service.uiState.value.copy(
+                    t9ResetSignal = service.uiState.value.t9ResetSignal + 1,
+                    t9RightCandidateSelectedCount = 0,
+                    t9SelectedCandidatePinyin = ""
+                )
+            }
+            // T9 跳过了 service.rimeEngine.commit()，需显式清除 RIME composition，
+            // 防止残留状态影响后续输入（在 service.keyProcessingDispatcher 上执行）。
+            if (isT9) service.rimeEngine.clearComposition()
+        } else {
+            withContext(Dispatchers.Main) {
+                if (isT9) {
+                    // partial commit：把本次选中的候选文本追加到累积列表，供后续合并显示
+                    if (selectedCandidate != null) {
+                        service.t9PartialCommitTexts.add(selectedCandidate)
+                    }
+                    // 保留状态字段，供 UI 层感知右侧选词事件
+                    service.uiState.value = service.uiState.value.copy(
+                        t9RightCandidateSelectedCount = service.uiState.value.t9RightCandidateSelectedCount + 1,
+                        t9SelectedCandidatePinyin = candidatePinyin ?: ""
+                    )
+                    // RIME composition 未被 selectCandidate 修改（已跳过），
+                    // 直接发送剩余数字到 RIME 重建 composition。
+                    service.keyboardCallbacks?.onT9ForceSendToRime?.invoke()
+                }
+                service.updateUI()
+            }
+        }
+    }
+    
+    /**
+     * 更新计算器候选栏显示
+     * 显示两个候选：
+     * - index 0: 计算结果（如 "2"），点击直接替换为结果
+     * - index 1: 带公式的结果（如 "1+1=2"），点击显示公式和结果
+     */
+    internal fun updateCalculatorCandidates() {
+        val candidate = service.calculatorEngine.getCandidate()
+        val result = service.calculatorEngine.getResult()
+        service.candidateState.value = if (candidate != null && result.isNotEmpty()) {
+            service.candidateState.value.copy(
+                candidates = listOf(result, candidate),
+                candidateComments = emptyList()
+            )
+        } else {
+            // 如果计算器之前有显示但现在已清除，也要清空候选栏
+            if (service.candidateState.value.candidates.isNotEmpty() && !service.calculatorEngine.isActive()) {
+                service.candidateState.value.copy(
+                    candidates = emptyList(),
+                    candidateComments = emptyList()
+                )
+            } else {
+                service.candidateState.value
+            }
+        }
+    }
+
+    internal fun selectCandidate(index: Int) {
+        service.composeViewRef?.let { service.feedbackManager.performKeyPressEffect(view = it) }
+
+        // 计算器模式（仅在数字/常用符号键盘下生效，防止状态残留导致其他键盘候选词点击异常）
+        val layoutState = service.keyboardViewModel.keyboardState.value
+        val isCalculatorKeyboard = layoutState is com.kingzcheung.xime.ui.keyboard.KeyboardLayoutState.Number
+                || layoutState is com.kingzcheung.xime.ui.keyboard.KeyboardLayoutState.CommonSymbol
+        if (service.calculatorEngine.isActive() && isCalculatorKeyboard) {
+            val result = service.calculatorEngine.getResult()
+            val expression = service.calculatorEngine.getExpression()
+            val formulaResult = service.calculatorEngine.getFormulaResult()
+            if (result.isNotEmpty() && expression.isNotEmpty()) {
+                val textToCommit: String
+                // index 0: 纯结果（如 "2"）
+                // index 1: 公式结果（如 "1+1=2"）
+                textToCommit = when (index) {
+                    0 -> result
+                    1 -> formulaResult
+                    else -> ""
+                }
+                if (textToCommit.isNotEmpty()) {
+                    service.calculatorEngine.clear()
+                    service.serviceScope.launch(Dispatchers.Main) {
+                        val ic = service.currentInputConnection
+                        if (ic != null) {
+                            // 删除输入框中已键入的表达式
+                            ic.deleteSurroundingText(expression.length, 0)
+                            // 提交选中的文本
+                            ic.commitText(textToCommit, textToCommit.length)
+                        }
+                        service.candidateState.value = CandidateState()
+                    }
+                }
+            }
+            return
+        }
+        
+        if (service.candidateState.value.isShowingRecentClipboard && index >= 0 && index < service.recentClipboardItemsState.value.size) {
+            val text = service.recentClipboardItemsState.value[index].text
+            service.textCommit.selectClipboardItem(text)
+            service.candidateState.value = service.candidateState.value.copy(
+                isShowingRecentClipboard = false,
+                candidates = emptyList(),
+                candidateComments = emptyList()
+            )
+        } else {
+            postRimeJob {
+                selectCandidateAsync(index)
+            }
+        }
+    }
+    
+    internal fun pageDown() {
+        postRimeJob {
+            service.rimeEngine.pageDown()
+            withContext(Dispatchers.Main) {
+                service.updateUI()
+            }
+        }
+    }
+    
+    internal fun pageUp() {
+        postRimeJob {
+            service.rimeEngine.pageUp()
+            withContext(Dispatchers.Main) {
+                service.updateUI()
+            }
+        }
+    }
+
+}
