+++ app/src/main/java/com/kingzcheung/xime/service/ImeTextCommit.kt
@@ -1,0 +1,112 @@
+package com.kingzcheung.xime.service
+
+import android.os.Build
+import android.util.Log
+import android.view.KeyEvent
+import android.view.inputmethod.InputConnection
+import android.view.inputmethod.InputContentInfo
+import androidx.core.content.FileProvider
+import java.io.File
+import java.io.FileInputStream
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+/**
+ * 文本上屏与剪贴板提交。
+ *
+ * 承载 commitImage（图片上屏）、剪贴板候选提交（selectClipboardItem/commitClipboardText/
+ * deleteClipboardChars）与语音撤销/搜索动作（performUndo/performSearch）。
+ * 共享状态通过 service 引用访问。
+ */
+internal class ImeTextCommit(private val service: XimeInputMethodService) {
+    internal fun performUndo() {
+        val currentTextBeforeCursor = service.currentInputConnection?.getTextBeforeCursor(1000, 0)?.toString() ?: ""
+        val currentLength = currentTextBeforeCursor.length
+        
+        val charsToDelete = currentLength - service.voiceRecognitionHandler.textLengthBeforeVoiceInput
+        
+        if (charsToDelete > 0) {
+            for (i in 0 until charsToDelete) {
+                service.currentInputConnection?.sendKeyEvent(KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_DEL))
+                service.currentInputConnection?.sendKeyEvent(KeyEvent(KeyEvent.ACTION_UP, KeyEvent.KEYCODE_DEL))
+            }
+        }
+        
+        service.voiceRecognitionHandler.textBeforeVoiceInput = ""
+        service.voiceRecognitionHandler.textLengthBeforeVoiceInput = 0
+    }
+    
+    internal fun performSearch() {
+        service.currentInputConnection?.sendKeyEvent(KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_ENTER))
+        service.currentInputConnection?.sendKeyEvent(KeyEvent(KeyEvent.ACTION_UP, KeyEvent.KEYCODE_ENTER))
+    }
+
+    internal fun commitImage(imagePath: String, mimeType: String = "image/jpeg"): Boolean {
+        return try {
+            val imageFile = File(imagePath)
+            if (!imageFile.exists()) {
+                Log.e(XimeInputMethodService.TAG, "Image file not found: $imagePath")
+                return false
+            }
+            
+            val cacheDir = File(service.cacheDir, "emoji_cache")
+            if (!service.cacheDir.exists()) {
+                service.cacheDir.mkdirs()
+            }
+            
+            val cacheFile = File(service.cacheDir, imageFile.name)
+            FileInputStream(imageFile).use { input ->
+                cacheFile.outputStream().use { output ->
+                    input.copyTo(output)
+                }
+            }
+            
+            val uri = FileProvider.getUriForFile(
+                service,
+                "$service.packageName.fileprovider",
+                cacheFile
+            )
+            
+            val inputContentInfo = InputContentInfo(
+                uri,
+                android.content.ClipDescription("emoji_image", arrayOf(mimeType)),
+                null
+            )
+            
+            val flags = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
+                InputConnection.INPUT_CONTENT_GRANT_READ_URI_PERMISSION
+            } else {
+                0
+            }
+            
+            service.currentInputConnection?.commitContent(inputContentInfo, flags, null) ?: false
+            
+        } catch (e: Exception) {
+            Log.e(XimeInputMethodService.TAG, "Failed to commit image", e)
+            false
+        }
+    }
+    
+
+    internal fun selectClipboardItem(text: String) {
+        if (service.candidateState.value.isComposing) {
+            service.keyRouter.postRimeJob {
+                service.rimeEngine.clearComposition()
+                withContext(Dispatchers.Main) {
+                    service.updateUI()
+                }
+            }
+        }
+        service.commitText(text)
+        service.clipboardManager.copyToSystemClipboard(text)
+    }
+
+    internal fun commitClipboardText(text: String) {
+        service.commitText(text)
+    }
+
+    internal fun deleteClipboardChars(count: Int) {
+        service.currentInputConnection?.deleteSurroundingText(count, 0)
+    }
+
+}
